Move tests from test_launch to test_main
This moves the tests for launch() to test_main, since launch() has been
moved to main.py.
